9C-1141-7 Rankings Processes
We extract from the processes module a `rankings` package.
